delete and add file
